fix auth file
