+++ server/controllers/middlewares/auth/index.ts
@@ -1,19 +1,3 @@
-import { RequestHandler } from 'express';
-import dotenv from 'dotenv';
-import { jwtVerify, CustomizedError } from '../../../utils';
+import checkAuth from './checkAuth';
 
-dotenv.config();
-
-const checkAuth: RequestHandler = async (req: any, res, next) => {
-  const { token } = req.cookies;
-  try {
-    if (!token) throw new CustomizedError(401, 'UnAuthorized');
-    const decoded = await jwtVerify(token);
-    req.user = decoded;
-    next();
-  } catch (err) {
-    next(new CustomizedError(401, 'UnAuthorized'));
-  }
-};
-
-export default checkAuth;
+export { checkAuth };
