+++ server/routes/index.ts
@@ -18,7 +18,7 @@ router.get('/categories', getCategories);
 router.get('/categories/:id/services', getCategoryServices);
 router.get('/book', getBooks);
 router.post('/services', postService);
-router.delete('/services/:id', deleteService);
+router.delete('/services/:id', checkAuth, deleteService);
 router.route('/contact').get(getContacts).post(addContact);
 router.post('/signup', signup);
 router.get('/logout', logout);
